Use virtual-core for common elements
