+++ package.json
@@ -39,7 +39,8 @@
   },
   "dependencies": {
     "@types/uuid": "^3.4.1",
-    "uuid": "^3.1.0"
+    "uuid": "^3.1.0",
+    "virtual-core": "0.0.8"
   },
   "devDependencies": {
     "@types/chai": "^4.0.4",
